Increase the camera far clip to the default of 1000 to avoid shadow issues
I noticed that shadows near the top of the game viewport were fading out. This was unexpected because the shadow distance was clearly long enough and the camera far clip was far enough to render the whole ground. I experimented and found that the camera far clip was causing the shadow fade out somehow. You can easily see this by checking out this commit and setting the far clip to 320. You'll see the gray cube in the top-right corner lose its shadow.